pyproject.toml
[project]
name = "streamlit-ai-resume-roast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.18.0",
    "ipykernel>=6.29.5",
    "ollama>=0.5.1",
    "openai>=1.82.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
